\name{plotl0}
\alias{plotl0}
\title{Plot L0 fitted value}
\description{
  This is a function that plots the fitted results of L0 method.
}

\usage{
 plotl0(re)
}
\arguments{
  \item{re}{The list generated by functions in FastSF package. For a given degree of freedom, plot the fitted value of L0 method. For adaptive situations, plot the MSE of every model under different degrees of freedom.}

}


\author{Canhong Wen, Xueqin Wang, Yanhe Shen, Aijun Zhang}



